Moved result-sort
