+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/result-sort/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './result-sort'
+export { default } from './container'
