+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/result-sort/presentation.tsx
@@ -16,10 +16,10 @@
 import { hot } from 'react-hot-loader'
 import * as React from 'react'
 import styled from 'styled-components'
-import MarionetteRegionContainer from '../../container/marionette-region-container'
-import { Button, buttonTypeEnum } from '../button'
+import MarionetteRegionContainer from '../container/marionette-region-container'
+import { Button, buttonTypeEnum } from '../presentation/button'
 
-const SortItemCollectionView = require('../../../component/sort/sort.view.js')
+const SortItemCollectionView = require('../../component/sort/sort.view.js')
 
 type Props = {
   removeSort: () => void
